requirements.txt
# Core
django==6.0.2
psycopg2-binary==2.9.11
python-decouple==3.8
whitenoise==6.8.0
gunicorn==23.0.0
pillow==11.1.0

# Auth
django-allauth[socialaccount]==65.3.0
requests==2.32.5          # People API calls (Google birthday fetch)

# Email
django-anymail[sendgrid]==12.0

# Async
celery[redis]==5.4.0
django-celery-beat==2.8.1
redis==5.2.1

# Push notifications
django-webpush==0.3.6

# Field-level encryption (OAuth secrets, PII)
django-encrypted-model-fields==0.6.5

# Rate limiting
django-ratelimit==4.1.0

# QR code generation
qrcode[pil]>=7.4

# Dev
django-debug-toolbar==6.2.0
pytest-django==4.9.0
pytest-cov==6.0.0
